Check for blocked countries during checkout

+++ ecommerce/views/v0/__init__.py
@@ -9,6 +9,7 @@ from main.constants import (
     USER_MSG_TYPE_PAYMENT_ERROR,
     USER_MSG_TYPE_PAYMENT_ERROR_UNKNOWN,
     USER_MSG_TYPE_PAYMENT_REVIEW,
+    USER_MSG_TYPE_ENROLL_BLOCKED,
 )
 from main.utils import redirect_with_user_message
 from rest_framework import mixins, status
@@ -415,7 +416,13 @@ class CheckoutCallbackView(View):
                 )
             )
             if basket:
-                basket.delete()
+                if basket.has_user_blocked_products(order.purchaser):
+                    return redirect_with_user_message(
+                        reverse("user-dashboard"),
+                        {"type": USER_MSG_TYPE_ENROLL_BLOCKED},
+                    )
+                else:
+                    basket.delete()
 
             return redirect_with_user_message(
                 reverse("user-dashboard"), {"type": USER_MSG_TYPE_PAYMENT_REVIEW}
@@ -465,9 +472,8 @@ class CheckoutInterstitialView(LoginRequiredMixin, TemplateView):
         except ObjectDoesNotExist:
             return HttpResponse("No basket")
 
-        if "no_checkout" in checkout_payload:
+        if "no_checkout" or "country_blocked" in checkout_payload:
             return checkout_payload["response"]
-
         return render(
             request,
             self.template_name,
